Move AttackGlobals to battle
