+++ battle/AttackGlobals.py
