+++ battle/Battle.py
@@ -7,7 +7,7 @@ from random import randint
 from typing import List, Tuple
 
 from battle.Attack import Attack
-from core.AttackGlobals import GROUP, MULTIPLIER, MULTIPLIER_DEFAULT
+from battle.AttackGlobals import GROUP, MULTIPLIER, MULTIPLIER_DEFAULT
 from core.Cog import Cog
 from core.Entity import BattleEntity, Entity
 from core.Exceptions import (CogAlreadyTrappedError, CogLuredError, Error,
